static memory class
adjusted the entirety of the memory class for static access.
Constructor method will populate both EEMemory and PS2BaseAddress.
Being that these 2 values will always be the same for the lifetime of the module , there is no need to obtain it consistently.

Tools class  was also adjusted to be a static class

all pointers for accessing members have been removed due to increased confusion